big Refactoring and fixes: move shared classes to shared submodule and fix serialization issues for transformation functions
