+++ etl/src/main/java/at/ac/uibk/dps/streamprocessingapplications/etl/FlinkJob.java
@@ -1,13 +1,22 @@
 package at.ac.uibk.dps.streamprocessingapplications.etl;
 
-import java.util.Arrays;
-import java.util.List;
+// spotless:off
+import at.ac.uibk.dps.streamprocessingapplications.etl.taxi.InterpolationFunction;
+import at.ac.uibk.dps.streamprocessingapplications.etl.taxi.RangeFilterFunction;
+import at.ac.uibk.dps.streamprocessingapplications.etl.transforms.ETLPipeline;
+import at.ac.uibk.dps.streamprocessingapplications.shared.TaxiSenMLParserJSON;
+import at.ac.uibk.dps.streamprocessingapplications.shared.model.TaxiRide;
 import org.apache.beam.runners.flink.FlinkPipelineOptions;
 import org.apache.beam.runners.flink.FlinkRunner;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.transforms.*;
+import org.apache.beam.sdk.values.TypeDescriptor;
+import org.apache.beam.sdk.values.TypeDescriptors;
 
+import java.util.stream.Stream;
+
+// spotless:on
 public class FlinkJob {
 
   public static void main(String[] args) {
@@ -16,53 +25,20 @@ public class FlinkJob {
     options.setRunner(FlinkRunner.class);
     // options.setParallelism(4);
 
-    List<String> sensorValues =
-        Arrays.asList(
-            "{\"bn\": \"temperature_sensor\", \"n\": \"temperature\", \"u\": \"Cel\","
-                + " \"v\": 25.3, \"t\": 1645467200}",
-            "{\"bn\": \"temperature_sensor\", \"n\": \"temperature\", \"u\": \"Cel\","
-                + " \"v\": 25.7, \"t\": 1645467500}",
-            "{\"bn\": \"temperature_sensor\", \"n\": \"temperature\", \"u\": \"Cel\","
-                + " \"v\": 26.2, \"t\": 1645467800}",
-            "{\"bn\": \"temperature_sensor\", \"n\": \"temperature\", \"u\": \"Cel\","
-                + " \"v\": 26.8, \"t\": 1645468100}",
-            "{\"bn\": \"temperature_sensor\", \"n\": \"temperature\", \"u\": \"Cel\","
-                + " \"v\": 27.1, \"t\": 1645468400}",
-            "{\"bn\": \"temperature_sensor\", \"n\": \"temperature\", \"u\": \"Cel\","
-                + " \"v\": 27.5, \"t\": 1645468700}",
-            "{\"bn\": \"temperature_sensor\", \"n\": \"temperature\", \"u\": \"Cel\","
-                + " \"v\": 27.9, \"t\": 1645469000}",
-            "{\"bn\": \"temperature_sensor\", \"n\": \"temperature\", \"u\": \"Cel\","
-                + " \"v\": 28.3, \"t\": 1645469300}",
-            "{\"bn\": \"temperature_sensor\", \"n\": \"temperature\", \"u\": \"Cel\","
-                + " \"v\": 28.7, \"t\": 1645469600}",
-            "{\"bn\": \"temperature_sensor\", \"n\": \"temperature\", \"u\": \"Cel\","
-                + " \"v\": 29.2, \"t\": 1645469900}",
-            "{\"bn\": \"humidity_sensor\", \"n\": \"humidity\", \"u\": \"%RH\", \"v\":"
-                + " 50, \"t\": 1645467200}",
-            "{\"bn\": \"humidity_sensor\", \"n\": \"humidity\", \"u\": \"%RH\", \"v\":"
-                + " 51, \"t\": 1645467500}",
-            "{\"bn\": \"humidity_sensor\", \"n\": \"humidity\", \"u\": \"%RH\", \"v\":"
-                + " 52, \"t\": 1645467800}",
-            "{\"bn\": \"humidity_sensor\", \"n\": \"humidity\", \"u\": \"%RH\", \"v\":"
-                + " 53, \"t\": 1645468100}",
-            "{\"bn\": \"humidity_sensor\", \"n\": \"humidity\", \"u\": \"%RH\", \"v\":"
-                + " 54, \"t\": 1645468400}",
-            "{\"bn\": \"humidity_sensor\", \"n\": \"humidity\", \"u\": \"%RH\", \"v\":"
-                + " 55, \"t\": 1645468700}",
-            "{\"bn\": \"humidity_sensor\", \"n\": \"humidity\", \"u\": \"%RH\", \"v\":"
-                + " 56, \"t\": 1645469000}",
-            "{\"bn\": \"humidity_sensor\", \"n\": \"humidity\", \"u\": \"%RH\", \"v\":"
-                + " 57, \"t\": 1645469300}",
-            "{\"bn\": \"humidity_sensor\", \"n\": \"humidity\", \"u\": \"%RH\", \"v\":"
-                + " 58, \"t\": 1645469600}",
-            "{\"bn\": \"humidity_sensor\", \"n\": \"humidity\", \"u\": \"%RH\", \"v\":"
-                + " 59, \"t\": 1645469900}");
-
     Pipeline pipeline = Pipeline.create(options);
 
+
     pipeline
-        .apply("Read data", Create.of(sensorValues))
+        .apply(Create.of(TaxiTestObjects.testPacks))
+        .apply(
+            new ETLPipeline<>(
+                TypeDescriptor.of(TaxiRide.class),
+                TaxiSenMLParserJSON::parseSenMLPack,
+                new RangeFilterFunction(),
+                TaxiTestObjects.buildTestBloomFilter(),
+                new InterpolationFunction(),
+                5))
+        .apply(MapElements.into(TypeDescriptors.strings()).via(TaxiRide::toString))
         .apply(ParDo.of(new FlinkJob.PrintFn()));
 
     pipeline.run();
