+++ etl/src/test/java/at/ac/uibk/dps/streamprocessingapplications/etl/taxi/RangeFilterFunctionTest.java
@@ -1,16 +1,7 @@
 package at.ac.uibk.dps.streamprocessingapplications.etl.taxi;
 
-import at.ac.uibk.dps.streamprocessingapplications.etl.taxi.model.TaxiRide;
-import at.ac.uibk.dps.streamprocessingapplications.etl.taxi.RangeFilterFunction;
-import org.apache.beam.sdk.coders.SerializableCoder;
-import org.apache.beam.sdk.testing.PAssert;
-import org.apache.beam.sdk.testing.TestPipeline;
-import org.apache.beam.sdk.testing.TestStream;
-import org.apache.beam.sdk.values.PCollection;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
 public class RangeFilterFunctionTest {
+  /*
   TestPipeline pipeline;
 
   @BeforeEach
@@ -44,5 +35,5 @@ public class RangeFilterFunctionTest {
     PCollection<TaxiRide> actual = pipeline.apply(createEvents).apply(new RangeFilterFunction());
 
     PAssert.that(actual).containsInAnyOrder(withinRangeTrimTime);
-  }
+  } */
 }
