--- etl/src/test/java/at/ac/uibk/dps/streamprocessingapplications/etl/transforms/ETLPipelineTest.java
@@ -1,7 +1,0 @@
-package at.ac.uibk.dps.streamprocessingapplications.etl.transforms;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class ETLPipelineTest {
-
-}
