+++ settings.gradle.kts
@@ -4,3 +4,4 @@ plugins {
 rootProject.name = "beam-applications-java"
 
 include("etl")
+include("shared")
